requirements.txt
# Python dependencies for EventIQ

# Web Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3  # Built-in with Python

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.2

# AI/ML Libraries
transformers==4.36.2
torch==2.1.2
openai==1.6.1
huggingface-hub==0.19.4

# PDF Generation
reportlab==4.0.8
weasyprint==60.2

# QR Code Generation
qrcode[pil]==7.4.2

# Image Processing
Pillow==10.1.0

# Email
sendgrid==6.11.0
emails==0.6

# Environment Management
python-decouple==3.8
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Logging
loguru==0.7.2

# Configuration
pyyaml==6.0.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Upload
python-multipart==0.0.6

# JSON Web Tokens
PyJWT==2.8.0

# Testing
factory-boy==3.3.0
faker==20.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.10

# PDF Generation for Certificates
reportlab==4.0.7
